Add files via upload
This is the old game in which I've made a year and half ago. At the time I've just leaned the basics about how to interact with the environment and how to make multiple systems work along. The code is simple and has around 520 lines of code of pure python